Splitting out reducers and actions
