+++ src/actions/dailygoal.js
@@ -1,0 +1,9 @@
+export const SET_DAILY_GOAL = "SET_DAILY_GOAL";
+
+export function setDailyGoal(day, text) {
+  return {
+    type: SET_DAILY_GOAL,
+    day: day,
+    text: text
+  };
+}
